Added create, update and delete forms with switch between them
