+++ implementation/sem-admin-client/src/components/CreateSubscriptionForm/ICreateFormFields.ts
@@ -1,0 +1,4 @@
+export interface ICreateFormFields {
+  url: string
+  enabledEvents: string[]
+}
