+++ implementation/sem-admin-client/src/components/DeleteSubscriptionForm/IDeleteFormFields.ts
@@ -1,0 +1,3 @@
+export interface IDeleteFormFields {
+  id: string
+}
