+++ implementation/sem-admin-client/src/components/UpdateSubscriptionForm/IUpdateFormFields.ts
@@ -1,0 +1,5 @@
+export interface IUpdateFormFields {
+  id: string
+  url: string
+  enabledEvents: string[]
+}
